Move contract store and governance stubs into services
