+++ docs/implementations/contract-store/collibra.md
@@ -31,8 +31,10 @@ adapter configuration so dc43 can route writes and reads consistently.
 ## Configuration example
 
 ```python
-from dc43.lib.components.contract_store import HttpCollibraContractAdapter
-from dc43.lib.components.contract_store.impl.collibra import CollibraContractStore
+from dc43.services.contracts.backend.stores import (
+    CollibraContractStore,
+    HttpCollibraContractAdapter,
+)
 
 adapter = HttpCollibraContractAdapter(
     base_url="https://collibra/api",
