+++ src/dc43/services/governance/backend/stubs/__init__.py
@@ -1,0 +1,5 @@
+"""Filesystem-backed sample implementations of the DQ governance interface."""
+
+from .filesystem import StubDQClient
+
+__all__ = ["StubDQClient"]
